moves repository to openchargemap package
